requirements.txt
boto3
pyyaml
pandas
openpyxl